+++ src/collective/chimpfeed/controlpanel.py
@@ -1,15 +1,21 @@
 import logging
 
 from Products.CMFCore.utils import getToolByName
-from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 
 from Acquisition import ImplicitAcquisitionWrapper
 
+from zope.app.pagetemplate.viewpagetemplatefile import ViewPageTemplateFile
 from zope.interface import implements
 from zope.component import adapts
 
 from plone.z3cform import layout
-from plone.app.registry.browser import controlpanel
+
+try:
+    from plone.app.registry.browser import controlpanel
+    switch_on = None
+except ImportError:
+    from collective.chimpfeed import legacy as controlpanel
+    from plone.z3cform.z2 import switch_on
 
 from collective.chimpfeed.interfaces import IFeedSettings
 from collective.chimpfeed.interfaces import IControlPanel
@@ -18,6 +24,7 @@ from collective.chimpfeed.feeds import make_urls
 
 from z3c.form import field
 from z3c.form import widget
+from z3c.form.browser import textlines
 from z3c.form.interfaces import IDataConverter
 
 logger = logging.getLogger("collective.chimpfeed.controlpanel")
@@ -56,6 +63,7 @@ class UrlsWidget(widget.Widget):
 class ControlPanelEditForm(controlpanel.RegistryEditForm):
     schema = IFeedSettings
     fields = field.Fields(IControlPanel)
+
     label = _(u"MailChimp settings")
     description = _(
         u"Please configure an API-key and define one or more "
@@ -65,6 +73,8 @@ class ControlPanelEditForm(controlpanel.RegistryEditForm):
     fields['urls'].mode = "display"
     fields['urls'].widgetFactory = UrlsWidget.factory
 
+    fields['feeds'].widgetFactory = textlines.TextLinesFieldWidget
+
     def updateActions(self):
         # This prevents a redirect to the main control panel page
         self.request.response.setStatus(200, lock=True)
@@ -72,6 +82,9 @@ class ControlPanelEditForm(controlpanel.RegistryEditForm):
         super(ControlPanelEditForm, self).updateActions()
 
     def render(self):
+        if switch_on is not None:
+            switch_on(self)
+
         if not self.getContent().feeds:
             del self.widgets['urls']
         else: